+++ SOS/Plush/TypeDef.h
@@ -12,18 +12,18 @@ namespace Plush
 		UNKNOWN = 5
 	};
 
-	template <typename T>  inline size_t get_stack_num() { /* typename_t_has_no_stack(); */ return UNKNOWN; }
+	template <typename T>  inline unsigned int get_stack_num() { /* typename_t_has_no_stack(); */ return UNKNOWN; }
 
-	template <> inline size_t get_stack_num<ExecAtom>() { return EXEC_STACK; }
-	template <> inline size_t get_stack_num<long>() { return INTEGER_STACK; }
-	template <> inline size_t get_stack_num<CodeAtom>() { return CODE_STACK; }
-	template <> inline size_t get_stack_num<bool>() { return BOOL_STACK; }
-	template <> inline size_t get_stack_num<double>() { return FLOAT_STACK; }
+	template <> inline unsigned int get_stack_num<ExecAtom>() { return EXEC_STACK; }
+	template <> inline unsigned int get_stack_num<long>() { return INTEGER_STACK; }
+	template <> inline unsigned int get_stack_num<CodeAtom>() { return CODE_STACK; }
+	template <> inline unsigned int get_stack_num<bool>() { return BOOL_STACK; }
+	template <> inline unsigned int get_stack_num<double>() { return FLOAT_STACK; }
 
 	template <typename T>
 	inline Type make_type() 
 	{
-		int n = get_stack_num<T>();
+		unsigned int n = get_stack_num<T>();
 		Type type(n + 1, n);
 		return type;
 	}
